+++ src/components/CombinedFeaturePanel/components/AccessibilitySection/OSMTagPanel.tsx
@@ -1,9 +1,5 @@
 import { get, set, sortBy } from "lodash";
 import * as React from "react";
-import {
-  getAvailableLangTags,
-  normalizeAndExtractLanguageTagsIfPresent,
-} from "~/components/CombinedFeaturePanel/utils/TagKeyUtils";
 import { useCurrentLanguageTagStrings } from "~/lib/context/LanguageTagContext";
 import type { TypeTaggedOSMFeature } from "~/lib/model/geo/AnyFeature";
 import { omittedKeyPrefixes } from "~/lib/model/osm/tag-config/omittedKeyPrefixes";
@@ -11,6 +7,10 @@ import { omittedKeySuffixes } from "~/lib/model/osm/tag-config/omittedKeySuffixe
 import { omittedKeys } from "~/lib/model/osm/tag-config/omittedKeys";
 import { pathsToConsumedTagKeys } from "~/lib/model/osm/tag-config/pathsToConsumedTagKeys";
 import { sortOrderMap } from "~/lib/model/osm/tag-config/sortOrderMap";
+import {
+  getAvailableLangTags,
+  normalizeAndExtractLanguageTagsIfPresent,
+} from "~/lib/util/TagKeyUtils";
 import isAccessibilityRelevantOSMKey from "../../../../lib/model/osm/tag-config/isAccessibilityRelevantOSMKey";
 import OSMTagTable from "./OSMTagTable";
