+++ src/domains/edit/components/AutoEditor.tsx
@@ -5,14 +5,14 @@ import React, { useContext, useState } from "react";
 import { toast } from "react-toastify";
 import useSWR, { mutate } from "swr";
 import { t } from "ttag";
-import { normalizeAndExtractLanguageTagsIfPresent } from "~/components/CombinedFeaturePanel/utils/TagKeyUtils";
+import { AppStateLink } from "~/components/App/AppStateLink";
+import { FeaturePanelContext } from "~/components/CombinedFeaturePanel/FeaturePanelContext";
+import { StyledReportView } from "~/components/CombinedFeaturePanel/ReportView";
 import { useEnvContext } from "~/lib/context/EnvContext";
 import { makeChangeRequestToInhouseApi } from "~/lib/fetchers/makeChangeRequestToInhouseApi";
 import { fetchFeaturePrefixedId } from "~/lib/fetchers/osm-api/fetchFeaturePrefixedId";
 import { isOSMFeature } from "~/lib/model/geo/AnyFeature";
-import { AppStateLink } from "~/components/App/AppStateLink";
-import { FeaturePanelContext } from "~/components/CombinedFeaturePanel/FeaturePanelContext";
-import { StyledReportView } from "~/components/CombinedFeaturePanel/ReportView";
+import { normalizeAndExtractLanguageTagsIfPresent } from "~/lib/util/TagKeyUtils";
 import FeatureNameHeader from "../../../components/CombinedFeaturePanel/components/FeatureNameHeader";
 import FeatureImage from "../../../components/CombinedFeaturePanel/components/image/FeatureImage";
 import getOsmParametersFromFeature from "../../../lib/fetchers/osm-api/getOsmParametersFromFeature";
